src/requirements.txt
mysql-connector-python==8.2.0
kafka-python==2.0.2
python-dotenv==1.0.0
